Move logger into logger/
