+++ unit_tests/test_splitter.py
@@ -5,9 +5,7 @@ simple_imports.import_distributor()
 
 from data_splitter import DataSplitter
 
-from logger import Logger
-from interprocess_communication import IPC
-
+# from logger.logger import Logger
 
 def execute_test(obj):
     for _ in range(10):
@@ -31,7 +29,7 @@ class RTGTest(unittest.TestCase):
         tools_active = execute_test(demo)
 
         for tool in tools_active:
-            self.assertTrue((tool.__class__ in (Logger, IPC)))
+            self.assertTrue((tool.__class__ is Logger))
 
     def test_use_ipc(self):
         """
